Code/requirements.txt
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.4
cymem==2.0.8
filelock==3.13.1
fsspec==2023.12.2
gensim==4.3.2
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy
opencv-python==4.7.0.68
packaging==23.2
pandas==1.5.3
pathlib_abc==0.1.1
pathy==0.11.0
Pillow==9.4.0
preshed==3.0.9
pydantic==1.10.13
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.4
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.2
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#sha256=65cea113d7509d5ab515583bbcbc9a31b4157bebbb73133d4110d2945476eaa4
torch
torchvision
tqdm==4.66.1
transformers==4.36.2
typer==0.7.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
matplotlib==3.9.3
